fix exception

+++ app/models.py
@@ -38,7 +38,7 @@ class Channel(Base):
     """
     __tablename__ = 'channel'
     id = Column(Integer, primary_key=True, index=True)
-    channel_id = Column(Integer, unique=True, index=True, nullable=True)  # The URL can come first but the channel ID populated later
+    channel_id = Column(BigInteger, unique=True, index=True, nullable=True)  # The URL can come first but the channel ID populated later
     channel_name = Column(String(256))
     channel_title = Column(String(256))
     channel_url = Column(String(256))
@@ -111,7 +111,7 @@ class Message(Base):
     message_id = Column(Integer, primary_key=True, index=True)
     chat_user_id = Column(BigInteger, ForeignKey('chat_user.chat_user_id'), nullable=False)
     account_id = Column(BigInteger, ForeignKey('account.account_id'), nullable=False)  # The account ID (bot)
-    channel_id = Column(Integer, ForeignKey('channel.channel_id'), nullable=False)
+    channel_id = Column(BigInteger, ForeignKey('channel.channel_id'), nullable=False)
     keyword_id = Column(Integer, ForeignKey('keyword.keyword_id'), nullable=False)
     message_text = Column(String(10000))
     message_is_mention = Column(Boolean, default=False)
@@ -158,7 +158,7 @@ class Notification(Base):
     id = Column(Integer, primary_key=True, index=True)
     keyword_id = Column(Integer, ForeignKey('keyword.keyword_id'), nullable=False)
     message_id = Column(Integer, ForeignKey('message.message_id'), nullable=False)
-    channel_id = Column(Integer, ForeignKey('channel.channel_id'), nullable=False)
+    channel_id = Column(BigInteger, ForeignKey('channel.channel_id'), nullable=False)
     account_id = Column(BigInteger, ForeignKey('account.account_id'), nullable=False)  # The account ID (bot)
     chat_user_id = Column(BigInteger, ForeignKey('chat_user.chat_user_id'), nullable=False)
     notification_tnotify = Column(DateTime, default=lambda: datetime.now(timezone.utc))
